store: Only the primary can look up the shard for a deployment
This change ensures that we only look up the deployment -> shard mapping in
the primary. In addition, since shards do not have any information about
the name of the schema that holds a deployment's data, pass the combination
of deployment id, shard name, and database namespace (schema) into lots of
places.